Extract db.patchCollection to define file
Cut-paste, no code has changed
